break down backup functionality into functions
