+++ api/src/backup/backup-service.ts
@@ -1,4 +1,4 @@
-import { Platform, User } from '@prisma/client';
+import { Backup, Platform, Playlist, User } from '@prisma/client';
 import prisma from '../util/prisma';
 
 export const createBackup = async (user: User, opts: {
@@ -8,7 +8,7 @@ export const createBackup = async (user: User, opts: {
     playlistName: string;
     playlistDescription: string;
     contentHash: string;
-    tracks: string[];
+    tracks: any[];
     followers: number;
     imageUrl: string;
     createdById: string;
@@ -24,6 +24,9 @@ export const createBackup = async (user: User, opts: {
             followers: opts.followers,
             tracks: opts.tracks,
             createdById: user.id
+        },
+        include: {
+            createdBy: true
         }
     })
 
@@ -34,9 +37,38 @@ export const createBackup = async (user: User, opts: {
             playlistId: playlist.id
         },
         include: {
+            playlist: true,
+            createdBy: true
+        }
+    });
+}
+
+export const generateManifest = async (mostRecentBackup: (Backup & { playlist: Playlist; }) | null, 
+                                            currentBackup: (Backup & { playlist: Playlist; }) | null) => {
+    // @ts-ignore
+    const diffAdded = currentBackup.playlist.tracks.filter(x => !mostRecentBackup?.playlist.tracks.includes(x));
+
+    // @ts-ignore
+    const diffRemoved = mostRecentBackup?.playlist.tracks.filter(x => !currentBackup.playlist.tracks.includes(x));
+
+    // @ts-ignore
+    const backup = await prisma.backup.update({
+        where: {
+            // @ts-ignore
+            id: currentBackup.id
+        },
+        data: {
+            manifest: {
+                added: diffAdded,
+                removed: diffRemoved
+            }
+        },
+        include: {
             playlist: true
         }
     });
+
+    return backup;
 }
 
 export const getMostRecentBackup = async (playlistId: string) => {
